Reference thesis has been changed.
I referred Yong Zeng's "Simultaneous Navigation and Radio Mapping
for Cellular-Connected UAV
with Deep Reinforcement Learning"

Due to the high complexity of rectifying tensorflow version, after discussion with my professor, a reference thesis has been changed.

To be specific, our goal is not to implement the simplest code, basic DQN and environment, but to implement slightly more difficult code that covers both DDDQN and environment with UAV.

The code available in Github would be rectified to add annotations to each line in order to utilize it as a basic code for future thesis writing.