Antenna Gain?
In the past, the experience of knowing that configuring an environment is the most important part of creating a code has led me to see the "radio_environment.ipynb" code first. The parts of constructing distribution of buildings were easily understood. However, when I met "defectAntennaGain(Theta_deg, Phi_deg):", I could not understand why this code was built in that way because I lacked basic antenna engineering knowledge. I could not fully understand other parts either due to concatenated effect of lack of getAntennaGain. So I am going to study about Antenna.