I've learned "radio_environment.ipynb"
After finding [5], [7] of the reference papers, I searched for antenna gain and finally found out how equations in def getAntennaGain came out. If there is sufficient time, it would be good to revisit the antenna detail model parameters or formulas in the [5], [7] of bibliographies.
Unfortunately, however, there are still many things to know about antenna modeling, such as array antenna, beamforming,rayian distribution, and rician distribution. Even more, due to the lack of physical time today, I haven't seen all of the environmental code.
So first of all, understanding the environmental code roughly, I felt that I have to study the algorithm that I was actually going to use in detail.