I think I need a GPU
The SNARM code was similar to the existing Duelling DDQN code, so it did not take much time to analyze.
After looking at all the codes, I organized them simply to understand them easily at a glance. (see 'code 정리.pdf')
After removing the sensorboard, there is no error in progressing the code, but it takes too much time to train.
For example, learning consisting of episode = 5000 takes two hours to progress only 3% of learning.
Therefore, I thought that I need to proceed with the gpu which is avaiable in google colab to learn practically.