state_action_reward
First of all, we tried to run the main algorithm using colab, but only learned about 15% in 6 hours.
To solve this problem, we need to take into account other ways, such as reducing the number of episodes or making the deep neural model lighter.

There don't seem to be as many papers registered in the journal or with some quotations as I thought.
Or even if the conditions are met, the paper is inaccessible due to some reasons.