Learning completed!
Originally, the number of episode was 5000, so the learning time was too long to see the exact result.
So, I could check the learning result by reducing the number of episodes to 200.
The results showed that the episode is only 200, and that the optimal value has not been found yet.
The results showed that the optimal value has not been found yet due to the lack of the number of episodes.