Multi-agent RL
In a paper related to multi-agent RL, I could see the environment and how state, action, and reward were defined.
Unfortunately, however, the paper's code was not available, making it difficult to understand exactly how it was implemented.
Therefore, it is necessary to find additional papers which release codes and analyze the codes as well as the environment, state, action, reward.