multi-agents code
It is too complicated for me to analyze the multi-agents code associated with communication. Therefore, It is reasonable to read other multi-agents code which is relatively simple than before ones. Although I didn't understand the simple code in depth, it seems that main difference between multi-agent and not multi-agent one is to define state space that would leads to change some functions like 'step' or 'action'. Unfortunately, however, I need a lot of time to fully understand multi-agent code.